+++ packages/partials/concurrent/src/Classes/MemoryMutex.test.ts
@@ -4,17 +4,49 @@ import { MemoryMutex, E_LOCK_FAILED } from './MemoryMutex';
 
 NodeTest.describe('Class MemoryMutex', async () => {
 
-    NodeTest.it('A new mutex should be able to lock and unlock', () => {
+    NodeTest.it('Option "reentrant" should be set correctly', () => {
+
+        NodeAssert.strictEqual(new MemoryMutex({ reentrant: true }).reentrant, true);
+        NodeAssert.strictEqual(new MemoryMutex({ reentrant: false }).reentrant, false);
+        NodeAssert.strictEqual(new MemoryMutex({}).reentrant, false);
+        NodeAssert.strictEqual(new MemoryMutex().reentrant, false);
+    });
+
+    NodeTest.it('New shared mutex should use the same reentrant option', () => {
+
+        NodeAssert.strictEqual(new MemoryMutex({ reentrant: true }).share().reentrant, true);
+        NodeAssert.strictEqual(new MemoryMutex({ reentrant: false }).share().reentrant, false);
+        NodeAssert.strictEqual(new MemoryMutex({}).share().reentrant, false);
+        NodeAssert.strictEqual(new MemoryMutex().share().reentrant, false);
+    });
+
+    NodeTest.it('A non-reentrant mutex should be able to lock and unlock', () => {
 
         const mutex = new MemoryMutex();
 
         NodeAssert.strictEqual(mutex.lock(), true);
         NodeAssert.strictEqual(mutex.isLocked(), true);
-        NodeAssert.strictEqual(mutex.lock(), true);
+        NodeAssert.strictEqual(mutex.lock(), false, 'Locking again should fail');
         NodeAssert.strictEqual(mutex.isLocked(), true);
         NodeAssert.strictEqual(mutex.unlock(), true);
         NodeAssert.strictEqual(mutex.isLocked(), false);
         NodeAssert.strictEqual(mutex.unlock(), false);
+        NodeAssert.strictEqual(mutex.unlock(), false);
+    });
+
+    NodeTest.it('A reentrant mutex should allow reacquiring the lock', () => {
+
+        const mutex = new MemoryMutex({ reentrant: true });
+
+        NodeAssert.strictEqual(mutex.lock(), true);
+        NodeAssert.strictEqual(mutex.isLocked(), true);
+        NodeAssert.strictEqual(mutex.lock(), true, 'Locking again should succeed');
+        NodeAssert.strictEqual(mutex.isLocked(), true);
+        NodeAssert.strictEqual(mutex.unlock(), true);
+        NodeAssert.strictEqual(mutex.unlock(), false);
+        NodeAssert.strictEqual(mutex.isLocked(), false);
+        NodeAssert.strictEqual(mutex.unlock(), false);
+        NodeAssert.strictEqual(mutex.isLocked(), false);
         NodeAssert.strictEqual(mutex.unlock(), false);
     });
 
@@ -27,8 +59,8 @@ NodeTest.describe('Class MemoryMutex', async () => {
         NodeAssert.strictEqual(mutex2.lock(), true);
         NodeAssert.strictEqual(mutex1.isLocked(), true);
         NodeAssert.strictEqual(mutex2.isLocked(), true);
-        NodeAssert.strictEqual(mutex1.lock(), true);
-        NodeAssert.strictEqual(mutex2.lock(), true);
+        NodeAssert.strictEqual(mutex1.lock(), false);
+        NodeAssert.strictEqual(mutex2.lock(), false);
         NodeAssert.strictEqual(mutex1.isLocked(), true);
         NodeAssert.strictEqual(mutex2.isLocked(), true);
         NodeAssert.strictEqual(mutex1.unlock(), true);
@@ -50,7 +82,7 @@ NodeTest.describe('Class MemoryMutex', async () => {
         NodeAssert.strictEqual(mutex2.lock(), false);
         NodeAssert.strictEqual(mutex1.isLocked(), true);
         NodeAssert.strictEqual(mutex2.isLocked(), false);
-        NodeAssert.strictEqual(mutex1.lock(), true);
+        NodeAssert.strictEqual(mutex1.lock(), false);
         NodeAssert.strictEqual(mutex2.lock(), false);
         NodeAssert.strictEqual(mutex1.isLocked(), true);
         NodeAssert.strictEqual(mutex2.isLocked(), false);
@@ -60,13 +92,11 @@ NodeTest.describe('Class MemoryMutex', async () => {
         NodeAssert.strictEqual(mutex2.isLocked(), false);
         NodeAssert.strictEqual(mutex2.unlock(), false);
         NodeAssert.strictEqual(mutex1.unlock(), false);
-        NodeAssert.strictEqual(mutex2.lock(), true);
-        NodeAssert.strictEqual(mutex1.isLocked(), false);
-        NodeAssert.strictEqual(mutex2.isLocked(), true);
-        NodeAssert.strictEqual(mutex1.unlock(), false);
-        NodeAssert.strictEqual(mutex2.unlock(), true);
-        NodeAssert.strictEqual(mutex1.isLocked(), false);
-        NodeAssert.strictEqual(mutex2.isLocked(), false);
+        NodeAssert.strictEqual(mutex2.unlock(), false);
+    });
+
+    NodeTest.it('Two shared reentrant mutex should share the mutex state', () => {
+
     });
 
     NodeTest.it('Method run should execute a synchronous function', () => {
@@ -151,6 +181,26 @@ NodeTest.describe('Class MemoryMutex', async () => {
         }, E_LOCK_FAILED);
 
         NodeAssert.strictEqual(mutex1.isLocked(), true);
+
+        NodeAssert.throws(() => {
+
+            mutex1.run(() => {
+
+                // Should not execute
+            });
+
+        }, E_LOCK_FAILED);
+
+        mutex1.unlock();
+
+        NodeAssert.strictEqual(mutex1.isLocked(), false);
+
+        NodeAssert.doesNotThrow(() => {
+            mutex1.run(() => {
+
+                // Now it can execute
+            });
+        });
     });
 
     NodeTest.it('Method run should return the result of the function', () => {